Refactor strages table in admininstration to use ViewComponents
